Changed intConnect parameter from int to void * Matches underlying devConnectInterruptVME() and allows us to pass 64 bit ptrs as the arg
git-svn-id: file:///afs/slac/g/pcds/vol2/svn/pcds/trunk/pcds/epics/modules/ipac/current@14667 ececa60f-9f63-0410-97c7-ea05888eb260

+++ drvTip810/drvTip810.c
@@ -885,7 +885,7 @@ int t810Initialise (
 	    status = errno;
 	}
 #else
-	if ((err=ipmIntConnect(pdevice->card, pdevice->slot, pdevice->irqNum, (void(*)())t810ISR, (int)pdevice))) {
+	if ((err=ipmIntConnect(pdevice->card, pdevice->slot, pdevice->irqNum, (void(*)())t810ISR, pdevice))) {
 		status = err;
 	}
 #endif
